refactor: move calculateDaysLeft to utils
