+++ src/components/GoalsComponents/MyGoal.tsx
@@ -8,7 +8,7 @@ import { unarchiveIcon } from "@src/assets";
 import { GoalItem } from "@src/models/GoalItem";
 import { unarchiveUserGoal } from "@src/api/GoalsAPI";
 import { replaceUrlsWithText, summarizeUrl } from "@src/utils/patterns";
-import { formatBudgetHrsToText } from "@src/utils";
+import { calculateDaysLeft, formatBudgetHrsToText } from "@src/utils";
 
 import { darkModeState, displayPartnerMode, lastAction } from "@src/store";
 import { displayGoalId, displayUpdateGoal, goalsHistory, displayChangesModal } from "@src/store/GoalsState";
@@ -43,23 +43,6 @@ const GoalSent = ({ goal }: { goal: GoalItem }) => {
 
   const hoursPerDayText = formatBudgetHrsToText(goal.timeBudget.perDay);
   const hoursPerWeekText = formatBudgetHrsToText(goal.timeBudget.perWeek);
-
-  const calculateDaysLeft = (dueDate: string) => {
-    if (!dueDate) return "";
-    const currentDate = new Date();
-    currentDate.setHours(0, 0, 0, 0);
-    const due = new Date(dueDate);
-    due.setHours(0, 0, 0, 0);
-
-    if (due.getTime() === currentDate.getTime()) {
-      return t("dueToday");
-    }
-
-    const timeDifference = due.getTime() - currentDate.getTime();
-    const daysLeft = Math.ceil(timeDifference / (1000 * 3600 * 24));
-
-    return daysLeft >= 0 ? t("daysLeft", { days: daysLeft }) : t("dueDatePassed");
-  };
 
   return (
     <>
